requirements.txt
numpy
pyyaml
sentence-transformers
scipy
matplotlib